djangoqr/requirement.txt
asgiref==3.11.0
colorama==0.4.6
Django==6.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pyzbar==0.1.9
qrcode==8.2
sqlparse==0.5.4
tzdata==2025.2
